Refactor headers
* Make color2.h, color4.h, vector2.h, vector4.h contain ONLY the
  structure defs, operator overloads, and the function overloads for
  functions already found in stdosl.h for the 3-component versions
  of the functions. Remove all other references to MaterialX and
  include directives from those files.

* Move all other {color,vector}{2,4} functions -- those specific to
  MaterialX functionality -- to func.h, grouped by function, so that
  for example all the remap functions are next to each other and can
  therefore have a single explanatory comment, and any further edits
  will all be adjacent.

* But for any of those sets of functions that are used solely in one
  mx shader, just move the function definitions into the .mx file and
  out of the headers that are read by all shaders.

* Rename func.h to mx_func.h.

* Remove exponent() and magnitude() functions in favor of pow and length,
  respectively, and remove a couple others that don't seem to have been
  used anywhere.
