+++ src/shaders/MaterialX/mx_exponent_float.mx
@@ -14,5 +14,5 @@ shader SHADER_NAME(mx_exponent_float)
     output TYPE out = TYPE_DEFAULT_OUT
   )
 {
-    out = exponent(in1, in2); 
+    out = pow(in1, in2);
 }
