+++ src/shaders/MaterialX/mx_hueshift.mx
@@ -6,6 +6,26 @@
 
 #include "mx_types.h"
 
+
+//
+// Shift the hue of a color by the given amount.
+//
+color hueshift(color in, float amount)
+{
+    color hsv3 = transformc("rgb","hsv", in);
+    hsv3[0] += amount;
+    hsv3[0] = fmod(hsv3[0], 1.0);
+    color out = transformc("hsv","rgb", hsv3);
+    return out;
+}
+
+color4 hueshift(color4 in, float amount)
+{
+    return color4 (hueshift(in.rgb, amount), in.a);
+}
+
+
+
 shader SHADER_NAME(mx_hueshift)
     [[ string help = "Shift the hue of a color; the alpha channel will be unchanged if present." ]]
   (
