+++ src/shaders/MaterialX/mx_matte.mx
@@ -6,6 +6,20 @@
 
 #include "mx_types.h"
 
+
+color2 cmatte(color2 fg, color2 bg)
+{
+    return color2((fg.r * fg.a) + bg.r * (1 - fg.a),
+                   fg.a + bg.a * (1 - fg.a));
+}
+
+color4 cmatte(color4 fg, color4 bg)
+{
+    return color4((fg.rgb * fg.a) + bg.rgb * (1 - fg.a),
+                   fg.a + bg.a * (1 - fg.a));
+}
+
+
 shader SHADER_NAME(mx_matte)
     [[ string help = TYPE_STR" Take two color2 or two color4 inputs and use the built-in alpha \
                      channel(s) to control the compositing of the fg and bg inputs: Ff+B(1-f)  (alpha: f+b(1-f))" ]]
