+++ src/shaders/MaterialX/mx_saturate.mx
@@ -6,6 +6,24 @@
 
 #include "mx_types.h"
 
+
+//
+// Shift the hue of a color by the given amount.
+//
+color saturate (color in, float amount)
+{
+    color hsv = transformc("rgb","hsv", in);
+    hsv[1] *= amount;
+    return transformc("hsv","rgb", hsv);
+}
+
+color4 saturate (color4 in, float amount)
+{
+    return color4 (saturate (in.rgb, amount), in.a);
+}
+
+
+
 shader SHADER_NAME(mx_saturate)
     [[ string help = "Adjust the saturation of a color; the alpha channel will be unchanged if present." ]]
   (
