+++ src/shaders/MaterialX/vector2.h
@@ -1,17 +1,22 @@
 // Open Shading Language : Copyright (c) 2009-2017 Sony Pictures Imageworks Inc., et al.
 // https://github.com/imageworks/OpenShadingLanguage/blob/master/LICENSE
-// 
-// MaterialX specification (c) 2017 Lucasfilm Ltd. 
-// http://www.materialx.org/
 
 #pragma once
-#include "funcs.h"
+#define VECTOR2_H
 
+
+// vector2 is a 2D vector
 struct vector2
 {
     float x;
     float y;
 };
+
+
+
+//
+// For vector2, define math operators to match vector
+//
 
 vector2 __operator__neg__(vector2 a)
 {
@@ -130,25 +135,26 @@ int __operator__ne__(vector2 a, vector2 b)
     return (a.x != b.x) || (a.y != b.y);
 }
 
+
+
+
+//
+// For vector2, define most of the stdosl functions to match vector
+//
+
 vector2 abs(vector2 a)
 {
-    return vector2 (abs(a.x),
-                    abs(a.y)
-                    );
+    return vector2 (abs(a.x), abs(a.y));
 }
 
 vector2 floor(vector2 a)
 {
-    return vector2 (floor(a.x),
-                    floor(a.y)
-                    );
+    return vector2 (floor(a.x), floor(a.y));
 }
 
 vector2 mix(vector2 a, vector2 b, float x )
 {
-    return vector2 (mix(a.x, b.x, x),
-                    mix(a.y, b.y, x)
-                    );
+    return vector2 (mix(a.x, b.x, x), mix(a.y, b.y, x));
 }
 
 float dot(vector2 a, vector2 b)
@@ -156,86 +162,32 @@ float dot(vector2 a, vector2 b)
     return (a.x * b.x + a.y * b.y);
 }
 
+float length (vector2 a)
+{
+    return hypot (a.x, a.y);
+}
+
 vector2 smoothstep(vector2 low, vector2 high, vector2 in)
 {
     return vector2 (smoothstep(low.x, high.x, in.x),
-                    smoothstep(low.y, high.y, in.y)
-                    );
+                    smoothstep(low.y, high.y, in.y));
 }
 
 vector2 smoothstep(float low, float high, vector2 in)
 {
     return vector2 (smoothstep(low, high, in.x),
-                    smoothstep(low, high, in.y)
-                    );
-}
-
-vector2 remap(vector2 in, vector2 inLow, vector2 inHigh, vector2 outLow, vector2 outHigh, int doClamp)
-{
-    //remap in from [inLow, inHigh] to [outLow, outHigh], optionally clamping to the new range
-
-    return vector2 (remap(in.x, inLow.x, inHigh.x, outLow.x, outHigh.x, doClamp),
-                    remap(in.y, inLow.y, inHigh.y, outLow.y, outHigh.y, doClamp)
-                    );
-}
-
-vector2 remap(vector2 in, float inLow, float inHigh, float outLow, float outHigh, int doClamp)
-{
-    //remap in from [inLow, inHigh] to [outLow, outHigh], optionally clamping to the new range
-
-    return vector2 (remap(in.x, inLow, inHigh, outLow, outHigh, doClamp),
-                    remap(in.y, inLow, inHigh, outLow, outHigh, doClamp)
-                    );
-}
-
-vector2 fgamma(vector2 in, vector2 g)
-{
-    return vector2 (fgamma(in.x, g.x),
-                    fgamma(in.y, g.y)
-                    );
-}
-
-vector2 fgamma(vector2 in, float g)
-{
-    return vector2 (fgamma(in.x, g),
-                    fgamma(in.y, g)
-                    );
+                    smoothstep(low, high, in.y));
 }
 
 vector2 clamp(vector2 in, vector2 low, vector2 high)
 {
     return vector2 (clamp(in.x, low.x, high.x),
-                    clamp(in.y, low.y, high.y)
-                    );
+                    clamp(in.y, low.y, high.y));
 }
 
 vector2 clamp(vector2 in, float low, float high)
 {
     return clamp(in, vector2(low, low), vector2(high, high));
-}
-
-vector2 contrast(vector2 in, vector2 amount, vector2 pivot)
-{
-    return vector2 (contrast(in.x, amount.x, pivot.x),
-                    contrast(in.y, amount.y, pivot.y)
-                    );
-}
-
-vector2 contrast(vector2 in, float amount, float pivot)
-{
-    return contrast(in, vector2(amount, amount), vector2(pivot, pivot));
-}
-
-vector2 exponent(vector2 in, vector2 amount)
-{
-    return vector2 (exponent(in.x, amount.x),
-                    exponent(in.y, amount.y)
-                    );
-}
-
-vector2 exponent(vector2 in, float amount)
-{
-    return exponent(in, vector2(amount, amount));
 }
 
 vector2 max(vector2 a, vector2 b)
@@ -251,15 +203,13 @@ vector2 max(vector2 a, float b)
 
 vector2 normalize(vector2 a)
 {
-    vector v = normalize(vector(a.x, a.y, 0));
-    return vector2 (v[0], v[1]);
+    return a / length(a);
 }
 
 vector2 min(vector2 a, vector2 b)
 {
     return vector2 (min(a.x, a.x),
-                    min(b.y, b.y)
-                    );
+                    min(b.y, b.y));
 }
 
 vector2 min(vector2 a, float b)
@@ -278,11 +228,6 @@ vector2 fmod(vector2 a, float b)
     return fmod(a, vector2(b, b));
 }
 
-float magnitude(vector2 a)
-{
-    return length(vector(a.x, a.y, 0));
-}
-
 vector2 rotate2d(vector2 in, float amount, vector2 center)
 {
     vector2 out = in - center;
@@ -294,3 +239,12 @@ vector2 rotate2d(vector2 in, float amount, vector2 center)
     return out;
 }
 
+vector2 pow (vector2 in, vector2 amount)
+{
+    return vector2 (pow(in.x, amount.x), pow(in.y, amount.y));
+}
+
+vector2 pow (vector2 in, float amount)
+{
+    return pow(in, vector2(amount, amount));
+}
